pyproject.toml
[project]
name = "restaurant-reservation-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.1",
    "django~=5.2",
    "django-celery-beat>=2.8.1",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt[crypto]>=5.5.1",
    "drf-spectacular>=0.29.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
]
